LPS-166870 Move additional methods to generic display context
